Rest Test
# Conflicts:
#	x-pack/plugin/inference/src/main/java/org/elasticsearch/xpack/inference/action/TransportCoordinatedInferenceAction.java
#	x-pack/plugin/ml/qa/ml-inference-service-tests/build.gradle
#	x-pack/plugin/ml/qa/ml-inference-service-tests/src/javaRestTest/java/org/elasticsearch/xpack/ml/integration/ExampleModels.java
