+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/inference/action/PutInferenceModelAction.java
@@ -5,61 +5,70 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.action;
+package org.elasticsearch.xpack.core.inference.action;
 
-import org.elasticsearch.TransportVersions;
 import org.elasticsearch.action.ActionRequestValidationException;
 import org.elasticsearch.action.ActionResponse;
 import org.elasticsearch.action.ActionType;
 import org.elasticsearch.action.support.master.AcknowledgedRequest;
+import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
+import org.elasticsearch.common.xcontent.XContentHelper;
 import org.elasticsearch.inference.ModelConfigurations;
 import org.elasticsearch.inference.TaskType;
 import org.elasticsearch.xcontent.ToXContentObject;
 import org.elasticsearch.xcontent.XContentBuilder;
+import org.elasticsearch.xcontent.XContentType;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
-public class GetInferenceModelAction extends ActionType<GetInferenceModelAction.Response> {
+public class PutInferenceModelAction extends ActionType<PutInferenceModelAction.Response> {
 
-    public static final GetInferenceModelAction INSTANCE = new GetInferenceModelAction();
-    public static final String NAME = "cluster:admin/xpack/inference/get";
+    public static final PutInferenceModelAction INSTANCE = new PutInferenceModelAction();
+    public static final String NAME = "cluster:admin/xpack/inference/put";
 
-    public GetInferenceModelAction() {
-        super(NAME, GetInferenceModelAction.Response::new);
+    public PutInferenceModelAction() {
+        super(NAME, PutInferenceModelAction.Response::new);
     }
 
-    public static class Request extends AcknowledgedRequest<GetInferenceModelAction.Request> {
+    public static class Request extends AcknowledgedRequest<Request> {
 
-        private final String modelId;
         private final TaskType taskType;
+        private final String modelId;
+        private final BytesReference content;
+        private final XContentType contentType;
 
-        public Request(String modelId, String taskType) {
-            this.modelId = modelId;
+        public Request(String taskType, String modelId, BytesReference content, XContentType contentType) {
             this.taskType = TaskType.fromStringOrStatusException(taskType);
+            this.modelId = modelId;
+            this.content = content;
+            this.contentType = contentType;
         }
 
         public Request(StreamInput in) throws IOException {
             super(in);
             this.modelId = in.readString();
             this.taskType = TaskType.fromStream(in);
+            this.content = in.readBytesReference();
+            this.contentType = in.readEnum(XContentType.class);
         }
 
-        @Override
-        public ActionRequestValidationException validate() {
-            return null;
+        public TaskType getTaskType() {
+            return taskType;
         }
 
         public String getModelId() {
             return modelId;
         }
 
-        public TaskType getTaskType() {
-            return taskType;
+        public BytesReference getContent() {
+            return content;
+        }
+
+        public XContentType getContentType() {
+            return contentType;
         }
 
         @Override
@@ -67,6 +76,13 @@ public class GetInferenceModelAction extends ActionType<GetInferenceModelAction.
             super.writeTo(out);
             out.writeString(modelId);
             taskType.writeTo(out);
+            out.writeBytesReference(content);
+            XContentHelper.writeTo(out, contentType);
+        }
+
+        @Override
+        public ActionRequestValidationException validate() {
+            return null;
         }
 
         @Override
@@ -74,71 +90,56 @@ public class GetInferenceModelAction extends ActionType<GetInferenceModelAction.
             if (this == o) return true;
             if (o == null || getClass() != o.getClass()) return false;
             Request request = (Request) o;
-            return Objects.equals(modelId, request.modelId) && taskType == request.taskType;
+            return taskType == request.taskType
+                && Objects.equals(modelId, request.modelId)
+                && Objects.equals(content, request.content)
+                && contentType == request.contentType;
         }
 
         @Override
         public int hashCode() {
-            return Objects.hash(modelId, taskType);
+            return Objects.hash(taskType, modelId, content, contentType);
         }
     }
 
     public static class Response extends ActionResponse implements ToXContentObject {
 
-        private final List<ModelConfigurations> models;
+        private final ModelConfigurations model;
 
-        public Response(List<ModelConfigurations> models) {
-            this.models = models;
+        public Response(ModelConfigurations model) {
+            this.model = model;
         }
 
         public Response(StreamInput in) throws IOException {
             super(in);
-            if (in.getTransportVersion().onOrAfter(TransportVersions.ML_INFERENCE_GET_MULTIPLE_MODELS)) {
-                models = in.readCollectionAsList(ModelConfigurations::new);
-            } else {
-                models = new ArrayList<>();
-                models.add(new ModelConfigurations(in));
-            }
+            model = new ModelConfigurations(in);
         }
 
-        public List<ModelConfigurations> getModels() {
-            return models;
+        public ModelConfigurations getModel() {
+            return model;
         }
 
         @Override
         public void writeTo(StreamOutput out) throws IOException {
-            if (out.getTransportVersion().onOrAfter(TransportVersions.ML_INFERENCE_GET_MULTIPLE_MODELS)) {
-                out.writeCollection(models);
-            } else {
-                models.get(0).writeTo(out);
-            }
+            model.writeTo(out);
         }
 
         @Override
         public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-            builder.startObject();
-            builder.startArray("models");
-            for (var model : models) {
-                if (model != null) {
-                    model.toXContent(builder, params);
-                }
-            }
-            builder.endArray();
-            builder.endObject();
-            return builder;
+            return model.toXContent(builder, params);
         }
 
         @Override
         public boolean equals(Object o) {
             if (this == o) return true;
             if (o == null || getClass() != o.getClass()) return false;
-            GetInferenceModelAction.Response response = (GetInferenceModelAction.Response) o;
-            return Objects.equals(models, response.models);
+            Response response = (Response) o;
+            return Objects.equals(model, response.model);
         }
 
         @Override
         public int hashCode() {
-            return Objects.hash(models);
+            return Objects.hash(model);
         }
     }
 }
