+++ x-pack/plugin/inference/src/internalClusterTest/java/org/elasticsearch/xpack/inference/integration/CoordinatedInferenceActionIT.java
@@ -12,9 +12,9 @@ import org.elasticsearch.core.Strings;
 import org.elasticsearch.plugins.Plugin;
 import org.elasticsearch.test.ESSingleNodeTestCase;
 import org.elasticsearch.xcontent.XContentType;
-import org.elasticsearch.xpack.core.inference.action.CoordinatedInferenceAction;
+import org.elasticsearch.xpack.core.ml.action.CoordinatedInferenceAction;
 import org.elasticsearch.xpack.inference.InferencePlugin;
-import org.elasticsearch.xpack.inference.action.PutInferenceModelAction;
+import org.elasticsearch.xpack.core.inference.action.PutInferenceModelAction;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
