+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/action/TransportCoordinatedInferenceAction.java
@@ -6,12 +6,10 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.action;
+package org.elasticsearch.xpack.ml.action;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.elasticsearch.ElasticsearchStatusException;
-import org.elasticsearch.ResourceNotFoundException;
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.support.ActionFilters;
 import org.elasticsearch.action.support.HandledTransportAction;
@@ -20,20 +18,18 @@ import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.util.concurrent.EsExecutors;
 import org.elasticsearch.inference.InferenceResults;
-import org.elasticsearch.rest.RestStatus;
+import org.elasticsearch.inference.TaskType;
 import org.elasticsearch.tasks.Task;
 import org.elasticsearch.transport.TransportService;
-import org.elasticsearch.xpack.core.inference.action.CoordinatedInferenceAction;
+import org.elasticsearch.xpack.core.inference.action.InferenceAction;
+import org.elasticsearch.xpack.core.ml.action.CoordinatedInferenceAction;
 import org.elasticsearch.xpack.core.ml.action.InferModelAction;
 import org.elasticsearch.xpack.core.ml.inference.assignment.TrainedModelAssignmentStateUtils;
 import org.elasticsearch.xpack.core.ml.inference.trainedmodel.EmptyConfigUpdate;
 import org.elasticsearch.xpack.core.ml.inference.trainedmodel.InferenceConfigUpdate;
-import org.elasticsearch.xpack.inference.UnparsedModel;
-import org.elasticsearch.xpack.inference.registry.ModelRegistry;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import static org.elasticsearch.xpack.core.ClientHelper.INFERENCE_ORIGIN;
 import static org.elasticsearch.xpack.core.ClientHelper.ML_ORIGIN;
@@ -46,7 +42,6 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
     private static final Logger logger = LogManager.getLogger(TransportCoordinatedInferenceAction.class);
 
     private final Client client;
-    private final ModelRegistry modelRegistry;
     private final ClusterService clusterService;
 
     @Inject
@@ -54,8 +49,7 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
         TransportService transportService,
         ActionFilters actionFilters,
         Client client,
-        ClusterService clusterService,
-        ModelRegistry modelRegistry
+        ClusterService clusterService
     ) {
         super(
             CoordinatedInferenceAction.NAME,
@@ -65,7 +59,6 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
             EsExecutors.DIRECT_EXECUTOR_SERVICE
         );
         this.client = client;
-        this.modelRegistry = modelRegistry;
         this.clusterService = clusterService;
     }
 
@@ -74,12 +67,13 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
         if (request.getModelType() == CoordinatedInferenceAction.Request.ModelType.NLP_MODEL) {
             // must be an inference service model or ml hosted model
             forNlp(request, listener);
+        } else if (request.hasObjects()) {
+            // Inference service models do not accept a document map
+            doInClusterModel(request, listener);  // if this fails check if IS model and error accordingly
         } else {
-            if (request.hasInferenceConfig() || request.hasObjects()) {
-                tryInClusterModel(request, listener);
-            } else {
-                tryInferenceServiceModel(request, listener);
-            }
+            forNlp(request, listener);  // if no model found check it isn't an NLP model
+
+            // boosted tree models need objects
         }
     }
 
@@ -115,37 +109,13 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
 
     private void doInferenceServiceModel(CoordinatedInferenceAction.Request request, ActionListener<InferModelAction.Response> listener) {
         logger.info("[CoordAction] doInferenceServiceModel [{}]", request.getModelId());
-        ActionListener<ModelRegistry.ModelConfigMap> modelMapListener = ActionListener.wrap(modelConfigMap -> {
-            var unparsedModel = UnparsedModel.unparsedModelFromMap(modelConfigMap.config(), modelConfigMap.secrets());
-            executeAsyncWithOrigin(
-                client,
-                INFERENCE_ORIGIN,
-                InferenceAction.INSTANCE,
-                new InferenceAction.Request(unparsedModel.taskType(), request.getModelId(), request.getInputs(), request.getTaskSettings()),
-                ActionListener.wrap(r -> translateInferenceServiceResponse(r.getResults(), listener), listener::onFailure)
-            );
-        }, listener::onFailure);
-
-        modelRegistry.getUnparsedModelMap(request.getModelId(), modelMapListener);
-    }
-
-    private void tryInferenceServiceModel(CoordinatedInferenceAction.Request request, ActionListener<InferModelAction.Response> listener) {
-        logger.info("[CoordAction] tryInferenceServiceModel [{}]", request.getModelId());
-        ActionListener<ModelRegistry.ModelConfigMap> modelMapListener = ActionListener.wrap(modelConfigMap -> {
-            logger.info("[CoordAction] tryInferenceServiceModel got model [{}]", "request.getModelId()");
-            var unparsedModel = UnparsedModel.unparsedModelFromMap(modelConfigMap.config(), modelConfigMap.secrets());
-            executeAsyncWithOrigin(
-                client,
-                INFERENCE_ORIGIN,
-                InferenceAction.INSTANCE,
-                new InferenceAction.Request(unparsedModel.taskType(), request.getModelId(), request.getInputs(), request.getTaskSettings()),
-                ActionListener.wrap(
-                    r -> translateInferenceServiceResponse(r.getResults(), listener),
-                    listener::onFailure
-                )
-            );
-
-        modelRegistry.getUnparsedModelMap(request.getModelId(), modelMapListener);
+        executeAsyncWithOrigin(
+            client,
+            INFERENCE_ORIGIN,
+            InferenceAction.INSTANCE,
+            new InferenceAction.Request(TaskType.ANY, request.getModelId(), request.getInputs(), request.getTaskSettings()),
+            ActionListener.wrap(r -> translateInferenceServiceResponse(r.getResults(), listener), listener::onFailure)
+        );
     }
 
     private void doInClusterModel(CoordinatedInferenceAction.Request request, ActionListener<InferModelAction.Response> listener) {
@@ -153,20 +123,6 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
         var inferModelRequest = translateRequest(request);
 
         executeAsyncWithOrigin(client, ML_ORIGIN, InferModelAction.INSTANCE, inferModelRequest, listener);
-    }
-
-    private void tryInClusterModel(CoordinatedInferenceAction.Request request, ActionListener<InferModelAction.Response> listener) {
-        logger.info("[CoordAction] tryInClusterModel [{}]", request.getModelId());
-        var inferModelRequest = translateRequest(request);
-
-
-        executeAsyncWithOrigin(
-            client,
-            ML_ORIGIN,
-            InferModelAction.INSTANCE,
-            inferModelRequest,
-            ActionListener.wrap(listener::onResponse, e -> handleDfaModelError(e, request, listener))
-        );
     }
 
     static InferModelAction.Request translateRequest(CoordinatedInferenceAction.Request request) {
@@ -180,7 +136,7 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
                 request.getObjectsToInfer(),
                 inferenceConfigUpdate,
                 request.getPreviouslyLicensed(),
-                request.getInferenceTimeout() // TODO is this timeout right
+                request.getInferenceTimeout()
             )
             : InferModelAction.Request.forTextInput(
                 request.getModelId(),
@@ -194,6 +150,23 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
         return inferModelRequest;
     }
 
+//    private ActionListener<InferModelAction.Response> wrapModelNotFoundInBoostedTreeModelCheck(ActionListener<InferModelAction.Response> listener) {
+////        return ActionListener.wrap(
+////            listener::onResponse,
+////            e -> {
+////                executeAsyncWithOrigin(
+////                    client,
+////                    ML_ORIGIN,
+////                    GetTrainedModelsAction.INSTANCE,
+////                    new InferenceAction.Request(TaskType.ANY, request.getModelId(), request.getInputs(), request.getTaskSettings()),
+////                    ActionListener.wrap(r -> translateInferenceServiceResponse(r.getResults(), listener), listener::onFailure)
+////                );
+////            }
+////        );
+//    }
+
+
+/*
     private void handleInferenceServiceModelFailure(
         Exception error,
         CoordinatedInferenceAction.Request request,
@@ -235,11 +208,19 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
         }
     }
 
-    private void lookForInferenceServiceModelWithId(String modelId, ActionListener<Optional<ModelRegistry.ModelConfigMap>> listener) {
-        logger.info("[CoordAction] lookForInferenceServiceModelWithId [{}]", modelId);
-        modelRegistry.getUnparsedModelMap(modelId, ActionListener.wrap(r -> listener.onResponse(Optional.of(r)), listener::onFailure));
-    }
 
+
+    private void lookForInferenceServiceModelWithId(String modelId, ActionListener<PutInferenceModelAction.Response> listener) {
+        logger.info("[CoordAction] lookForInferenceServiceModelWithId [{}]", modelId);
+        executeAsyncWithOrigin(
+            client,
+            INFERENCE_ORIGIN,
+            GetInferenceModelAction.INSTANCE,
+            new GetInferenceModelAction.Request(modelId, TaskType.ANY),
+            listener
+        );
+    }
+*/
     static void translateInferenceServiceResponse(
         List<? extends InferenceResults> inferenceResults,
         ActionListener<InferModelAction.Response> listener
@@ -247,13 +228,8 @@ public class TransportCoordinatedInferenceAction extends HandledTransportAction<
         logger.info("[CoordAction] translateInferenceServiceResponse");
         assert inferenceResults.size() > 0 : "no results";
 
-        try {
-            var copyListDueToBoundedCaptureError = new ArrayList<InferenceResults>(inferenceResults);
-            var cp = new InferModelAction.Response(copyListDueToBoundedCaptureError, null, false);
-            listener.onResponse(cp);
-        } catch (Exception e) {
-            logger.error(e);
-            listener.onFailure(e);
-        }
+        var copyListDueToBoundedCaptureError = new ArrayList<InferenceResults>(inferenceResults);
+        var cp = new InferModelAction.Response(copyListDueToBoundedCaptureError, null, false);
+        listener.onResponse(cp);
     }
 }
